6langchain_rag/requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.4.1
langchain-openai==0.2.2
langchain-text-splitters==0.3.2

qdrant-client==1.12.1

tiktoken==0.7.0
python-dotenv==1.0.1

